parser.ExplicitFiles2ModulesFile: create label definitions in the ModulesFile for all labels in .lab file
To be able to reference labels that are loaded from a .lab file and stored in the model,
they have to appear in the ModulesFile. This method stores

label "labelName" = false;

definitions. As the actual state set information is directly stored in the model,
evaluating the label will result in the correct state set.