ODD: Use int64_t for offsets everywhere, check for arithmetic overflow
Before, the ODD offsets (for indexing into the reachable states) were
stored as longs, resulting in inconsistent numbers of (symbolic)
states that can be treated. In particular, on Win64, long is 32bit.

We switch to int64_t for the ODD offsets everywhere, allowing
MTBDD computations for models with >2^31 states.

Additionally, we check for arithmetic overflow in the offset
computations.  Due to the symbolic storage for the models and the
state space explosion, it's possible to construct simple model files
that reach the 2^63 states limit.

Currently, this leads to an error message. In the future it might be
worthwile to ensure that PRISM can deal with the absence of an ODD for
all the computations that don't absolutely require it and carry on
(most MTBDD engine algorithms should be fine).