Sparse engine: Error if number of reachable states is too large
Currently, the sparse engine internally uses int (signed 32bit) index variables
So, if the number of states is larger than Integer.MAX_VALUE, there is a problem
and the code will most probably crash or do nonsensical things.

We check this before calling into the native sparse engine code and
throw a PrismNotSupportedException.