Relax restrictions on global variable updates.
Before, no global variables in PRISM models could be updated in a synchronizing (i.e., with an action name)
action. This patch relaxes this restriction. We now allow updates in synchronizing actions if we can match
each global variable to at most one module that performs updates to this variable in synchronizing actions.

For the explicit and simulator engines, no change is needed to support this apart from the syntactic checking.

For the symbolic engines, we have to track the updates to global variables and handle synchronizing actions
in a smart way.