Makefile: statically link pthread library for cygwin
In recent mingw g++ compilers from cygwin, the standard C++ library
depends on the pthreads library (libwinpthread-1.dll). As we'd like
for the prism DLLs to not depend on that library, we adapt the
linker flags to statically link the pthread library.

As we are not actually using any symbols, we also have to use the
--whole-archive flag to force linking.

https://stackoverflow.com/questions/13768515/how-to-do-static-linking-of-libwinpthread-1-dll-in-mingw/43402308#43402308