symbolic LTLModelChecker: convert LTL product to use statesOfInterest
This starts the reachability in the product from the set of states of interest
(with the corresponding initial state in the deterministic automaton).

Additionally, these statesOfInterest in the product become the
initial states of the product model.

Deprecated legacy methods for obtaining product models as before (with
the initial states in the product model being those that were initial states
in the original model) are provided.