param: Print warning if heuristic checking was used
To check that, for all parameter valuations in a region,
(1) some threshold property holds holds or (2) that a given
scheduler is optimal, a set of constraints is checked.

Currently, only heuristic checking is implemented, i.e., to determine
via sampling if the constraints hold. This can produce false positives
in the region computations, i.e., a region is not sufficiently split.

We add a warning if the region computation relied on the heuristic
constraint check and thus can be imprecise.