parametric / exact engine: fix Rmin[F] scheduler precomputation for policy iteration
To ensure that the policy iteration (performed on an MDP where parameters have been replaced
by a parameter valuation) converges and converges on the right result, we initialise the
policy iteration for Rmin[F] with a proper scheduler, i.e., that reaches the goal
with probability one (see [BertsekasTsitsiklis91]).

Together with the previous commit, this fixes issue
https://github.com/prismmodelchecker/prism/issues/15.

As there is no check against negative rewards, those remain problematic.