parametric engine: fix handling of infinite values in reward computations
In the state eliminator, i.e., solving for a DTMC, astate that has probability
zero of reaching the target states (i.e., that can not reach the target state)
should get infinite reward.

Previously, the check for this looked at the returned set of collectStatesBackward(),
which always returns the whole state space (used for backward elimination order).

Now, we use the variant of collectStatesBackward() that only returns the states
that can reach the target set.

Additionally, we now compute the set of infinity states for Rmin/Rmax using
prob0a / prob0e, respectively, and ensure that they get value infinity.

This fixes